refactor: move supported languages to code block
also handle highlightedLines there
