+++ packages/ui/src/components/CodeBlock/CodeBlock.tsx
@@ -1,9 +1,9 @@
 import { ConfiguredRefractor } from './ConfiguredRefractor';
 
 interface ICodeBlockProps {
-  code?: string;
+  code: string;
   language?: string;
-  highlightedLines?: number[];
+  highlightedLines?: string;
   filename?: string;
 }
 
@@ -11,7 +11,7 @@ export function CodeBlock({
   code,
   language,
   filename,
-  highlightedLines = [],
+  highlightedLines,
 }: ICodeBlockProps) {
   if (!code) {
     return null;
@@ -29,7 +29,7 @@ export function CodeBlock({
       <ConfiguredRefractor
         language={lang}
         value={code}
-        markers={highlightedLines}
+        markers={getLines(highlightedLines)}
       />
     </div>
   );
@@ -41,4 +41,30 @@ function Filename({ filename }: { filename: string }) {
       <span>{filename}</span>
     </div>
   );
+}
+
+function getLines(lines?: string): number[] {
+  if (!lines || !lines.trim()) {
+    return [];
+  }
+  const allLines = lines.split(',');
+
+  const result = [];
+
+  for (const line of allLines) {
+    const [initial, end] = line.split('-');
+    const initialParsed = parseInt(initial, 10);
+
+    if (!end) {
+      result.push(initialParsed);
+    }
+
+    const endParsed = parseInt(end, 10);
+
+    for (let i = initialParsed; i <= endParsed; i++) {
+      result.push(i);
+    }
+  }
+
+  return result;
 }
