+++ packages/ui/src/components/CodeBlock/index.ts
@@ -1,1 +1,2 @@
 export * from './CodeBlock';
+export * from './supportedLanguages';
